refactor: split parsing service internals
Move parsing state, ingest orchestration, and batch execution into focused modules while keeping ParsingService and its test-visible patch surfaces stable.
